+++ src/canvas/components/data_table/styling.rs
@@ -11,6 +11,7 @@ pub struct DataTableStyling {
     pub text_style: Style,
     pub highlighted_text_style: Style,
     pub title_style: Style,
+    pub general_widget_style: Style,
 }
 
 impl DataTableStyling {
@@ -23,6 +24,7 @@ impl DataTableStyling {
             text_style: styles.text_style,
             highlighted_text_style: styles.selected_text_style,
             title_style: styles.widget_title_style,
+            general_widget_style: styles.general_widget_style,
         }
     }
 }