+++ src/options/config/style/widgets.rs
@@ -33,4 +33,7 @@ pub(crate) struct WidgetStyle {
 
     /// Widget borders type.
     pub(crate) widget_border_type: Option<WidgetBorderType>,
+
+    /// Background color for widgets.
+    pub(crate) bg_color: Option<ColorStr>,
 }